More Migrations to TS (#720)
* Migrate constants to TS
* Change type of Rarity; exclude Land from CARD_RARITY loop
* Some bugfixes
* Fix display of needed cards sprite
* npm run format; remove commented out code
* Rename colors.js
* Fix missing types; fix import directives
* Refactorings
