+++ src/shared/CardTile.tsx
@@ -15,7 +15,7 @@ import {
   openScryfallCard
 } from "./util";
 import { addCardHover } from "./card-hover";
-import { DbCardData } from "./types/Metadata";
+import { DbCardData, Rarity } from "./types/Metadata";
 
 export interface CardTileProps {
   card: DbCardData | any; // TODO remove group lands hack
@@ -128,33 +128,6 @@ function CostSymbols(props: { card: DbCardData; dfcCard?: DbCardData }): JSX.Ele
   return <>{costSymbols}</>;
 }
 
-function ArenaWildcardsNeeded(props: {
-  card: DbCardData;
-  deck: Deck;
-  isSideboard: boolean;
-  ww: number;
-}): JSX.Element {
-  const { card, deck, isSideboard, ww } = props;
-  if (card.type.indexOf("Basic Land") === -1) {
-    const missing = getWildcardsMissing(deck, card.id, isSideboard);
-    if (missing > 0) {
-      const xoff = CARD_RARITIES.indexOf(card.rarity) * -24;
-      const yoff = missing * -24;
-      return (
-        <div
-          className="not_owned_sprite"
-          title={missing + " missing"}
-          style={{
-            backgroundPosition: `${xoff}px ${yoff}px`,
-            left: `calc(0px - 100% + ${ww - 14}px)`
-          }}
-        />
-      );
-    }
-  }
-  return <></>;
-}
-
 function ArenaCardTile(props: CardTileProps): JSX.Element {
   const {
     card,
@@ -238,10 +211,11 @@ function ArenaCardTile(props: CardTileProps): JSX.Element {
         }}
       />
       {showWildcards && deck && (
-        <ArenaWildcardsNeeded
+        <WildcardsNeeded
           card={card}
           deck={deck}
           isSideboard={isSideboard}
+          listStyle="arena"
           ww={ww}
         />
       )}
@@ -272,29 +246,59 @@ function FlatQuantityDisplay(props: { quantity: any }): JSX.Element {
   }
 }
 
-function FlatWildcardsNeeded(props: {
-  card: DbCardData;
-  deck: Deck;
-  isSideboard: boolean;
-}): JSX.Element {
-  const { card, deck, isSideboard } = props;
+interface WildcardsNeededProps{
+  card:DbCardData;
+  deck:Deck;
+  isSideboard:boolean;
+  listStyle:"flat"|"arena";
+  ww?:number;
+}
+
+interface MissingCardsProps{
+  missing:number;
+  cardRarity:Rarity;
+  listStyle:"flat"|"arena";
+  ww?:number;
+}
+
+function WildcardsNeeded(props: WildcardsNeededProps): JSX.Element {
+  const { card, deck, isSideboard, listStyle, ww } = props;
   if (card.type.indexOf("Basic Land") === -1) {
     const missing = getWildcardsMissing(deck, card.id, isSideboard);
+    const cardRarity = card.rarity;
+
+
     if (missing > 0) {
-      const xoff = CARD_RARITIES.indexOf(card.rarity) * -24;
-      const yoff = missing * -24;
-      return (
-        <div
-          className="not_owned_sprite_flat"
-          title={missing + " missing"}
-          style={{
-            backgroundPosition: `${xoff}px ${yoff}px`
-          }}
-        />
-      );
+      return MissingCardSprite({missing, cardRarity, listStyle, ww});
     }
   }
   return <></>;
+}
+
+function MissingCardSprite(props:MissingCardsProps):JSX.Element{
+  const{missing, cardRarity, listStyle, ww} = props;
+
+  const xoff = CARD_RARITIES.indexOf(cardRarity) * -24;
+  const yoff = missing * -24;
+
+  var className = "not_owned_sprite";
+  if(listStyle === "flat"){
+    className += "_flat";
+  }
+
+  const style:React.CSSProperties =
+  {backgroundPosition: `${xoff}px ${yoff}px`};
+  if(ww){
+    style.left = `calc(0px - 100% + ${ww - 14}px)`;
+  }
+
+  return (
+    <div
+    className={className}
+    title={missing + " missing"}
+    style={style}
+    />
+    );
 }
 
 function FlatCardTile(props: CardTileProps): JSX.Element {
@@ -391,10 +395,11 @@ function FlatCardTile(props: CardTileProps): JSX.Element {
         <CostSymbols card={card} dfcCard={dfcCard} />
       </div>
       {showWildcards && deck && (
-        <FlatWildcardsNeeded
+        <WildcardsNeeded
           card={card}
           deck={deck}
           isSideboard={isSideboard}
+          listStyle = "flat"
         />
       )}
     </div>
