+++ src/window_main/decks.js
@@ -179,7 +179,8 @@ export function openDecksTab(_filters = {}, scrollTop = 0) {
     let wc;
     let n = 0;
     let boosterCost = getBoosterCountEstimate(missingWildcards);
-    CARD_RARITIES.forEach(cardRarity => {
+    CARD_RARITIES.filter(rarity => rarity !== "Land").forEach(cardRarity => {
+      cardRarity = cardRarity.toLowerCase();
       if (missingWildcards[cardRarity]) {
         n++;
         wc = createDiv(["wc_explore_cost", "wc_" + cardRarity]);
