+++ src/window_main/explore.js
@@ -564,7 +564,8 @@ function deckLoad(_deck, index) {
   let wc;
   let n = 0;
   let boosterCost = getBoosterCountEstimate(_deck.wildcards);
-  CARD_RARITIES.forEach(rarity => {
+  CARD_RARITIES.filter(r => r !== "Land").forEach(rarity => {
+    rarity = rarity.toLowerCase();
     const key = rarity[0];
     if (_deck.wildcards.hasOwnProperty(key) && _deck.wildcards[key] > 0) {
       wc = createDiv(["wc_explore_cost", "wc_" + rarity], _deck.wildcards[key]);
